requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
dj-database-url==2.1.0
Django
django-crispy-forms==2.0
et-xmlfile==1.1.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
protobuf==4.24.1
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
psycopg2==2.9.6
whitenoise==6.4.0
gunicorn==20.1.0
djangorestframework==3.14.0
python-telegram-bot
